+++ src/profile.html
@@ -61,18 +61,15 @@
                             <p id="user_bio" class="text-center mr-3 ml-3">User bio section, where user can write a short message about themselves or their page.</p>
                         </div>
                         <div id="follow_counter_div" class="container-fluid pt-2 pb-0 text-center">
-                            <span id="follow_counter_badge" class="badge badge-pill badge-info"><span id="followerCounter">0</span> Followers</span>
+                            <span id="follow_counter_badge" class="badge badge-pill badge-info">000 Followers</span>
                         </div>
                         <div class="container-fluid text-center mt-4">
-                            <button id="follow_user" type="button" class="btn btn-outline-primary border mb-2">Follow</button>
+                            <button id="follow_user" type="button" class="btn btn-primary mb-2">FOLLOW</button>
                         </div>
 
                     </div>
                 </div>
             </div>
-        </div>
-
-        <div>
         </div>
 
         <script src="static/js/tether.min.js"></script>
@@ -82,7 +79,6 @@
         <script src="static/js/tweet.js"></script>
         <script src="static/js/tweet_functions.js"></script>
         <script src="static/js/profile.js"></script>
-        <script src="static/js/serverResponses.js"></script>
 
     </body>
