--- src/static/js/serverResponses.js
@@ -1,6 +1,0 @@
-var ServerResponses = {
-    SUCCESS: 'success',
-    INVALID_CREDENTIALS: 'invalid user or email',
-    OTHER: "other",
-    FAILURE: 'fail'
-};
